refactor tests to use shared helpers
